Create odd_or_even
simple logic to determine if a number is divisible by 2.  Additional logic to determine if a number is divisible by 4 and 2.  Could have designed a better if/else logic, eg. if num % 2 == 0 and num % 4 == 0: print "this number is divisible by 4.  Seeing that now looking back on the code.  Did this a while ago when I was just starting.